+++ moli-renderer-v8/src/host/timers.rs
@@ -266,7 +266,7 @@ impl HostTimeoutScheduler {
         &mut self,
         scope: &mut v8::PinScope<'s, '_>,
         callback: v8::Local<'s, v8::Function>,
-        delay_ms: u32,
+        delay_ms: u64,
         owner: HostTimerOwner,
         extra_args: Vec<v8::Global<v8::Value>>,
     ) -> u32 {
@@ -324,7 +324,7 @@ impl HostTimeoutScheduler {
                     is_interval: false,
                     extra_args,
                 },
-                delay_ms,
+                u64::from(delay_ms),
                 Instant::now(),
             )
             .get()
@@ -477,7 +477,7 @@ impl HostTimeoutScheduler {
                     is_interval,
                     extra_args,
                 },
-                delay_ms,
+                u64::from(delay_ms),
                 Instant::now(),
             )
             .get()
@@ -529,7 +529,7 @@ impl HostTimeoutScheduler {
                     is_interval: false,
                     extra_args,
                 },
-                delay_ms,
+                u64::from(delay_ms),
                 Instant::now(),
             )
             .get()
@@ -581,7 +581,7 @@ impl HostTimeoutScheduler {
                     is_interval: true,
                     extra_args,
                 },
-                delay_ms,
+                u64::from(delay_ms),
                 Instant::now(),
             )
             .get()
