+++ moli-renderer-v8/src/native_bridge/context_host/host_loads.rs
@@ -645,7 +645,7 @@ impl JsContextHost {
         let timer_id = self.queue_timeout(
             scope,
             callback,
-            navigation.delay_ms,
+            u64::from(navigation.delay_ms),
             HostTimerOwner::ChildWindow(handle),
             Vec::new(),
         );
